+++ src/pages/about.tsx
@@ -1,6 +1,6 @@
-import { Content } from '../components/Content';
 import DevIcons from '../components/DevIcons';
-import { Main } from '../components/Main';
+import { Content } from '../layout/Content';
+import { Main } from '../layout/Main';
 import { Meta } from '../layout/Meta';
 
 const About = () => (
